backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
faiss-cpu==1.7.4
pymupdf==1.23.7
sentence-transformers==2.3.1
openai==1.6.1
tiktoken==0.5.2
chromadb==0.4.22
pillow==10.2.0
gunicorn==21.2.0
dj-database-url==2.1.0
